Implement AsyncFileSystem for PseudoFs
In preparation for adding `async_readdir` tgo the AsyncFileSystem trait,
this implements `AsyncFileSystem` for `PseudoFs` so it will be easier to
implement `async_readdir` for `PseudoFs`. Prior to this, async `Vfs`
could use the sync version of `PseudoFs`, however, due to how writing
dir entries works, the async `Vfs` cannot use the sync version of
`PseudoFs::readdir`.

Alternatively, sync readdir could return an iterator which would make it
easier to convert to an `Stream`, but that would be a breaking change to
the sync interface.